Added quite a bit to the run_sql_query(),_existing_tables(),check_query_status(),check_all_jobs(),completed_job_info(),delete_job(),and delete_all_jobs() functions, in addition to adding some preliminary docstrings for many of them.
Did quite a bit more work on a number of the functions I worked on last night. Many of them are now fairly stable, with the exception of deleting jobs with the phase EXECUTING. Next should be handling the actual data stored in the tables for COMPLETED jobs.

Deleted the MultiDark directory and replaced it with the CosmoSim directory. Added tests directory with empty files for testing purposes.

Minor changes to the core.py file function download().