Deleted the MultiDark directory and replaced it with the CosmoSim directory. Added tests directory with empty files for testing purposes.
Minor changes to the core.py file function download().

Got the core functionality of download() working. The function now parses a completed table response and either returns the data to the terminal, or it writes to file and returns the data to the terminal.

Minor changes

Minor changes.

Layed out the framework for the MultiDark query tool. This is the first commit.

Added beginnings of query functions in MultiDark.

Added blank test_multidark_remote.py file to tests directory. This will test wheter a login is successful or not.

Added some basic structure and code to the multidark folder.

Adding files.

minor fixes..

Layed out the framework for the MultiDark query tool. This is the first commit.

Added beginnings of query functions in MultiDark.

Added blank test_multidark_remote.py file to tests directory. This will test wheter a login is successful or not.

Added some basic structure and code to the multidark folder.

Adding files.

minor fixes..

minor changes

Added cosmosim repository, which will contain the machinery required to query the MultiDark, BolshoiP, and CLUES cosmological simulation databases. The multidark repository will become deprecated.

First cosmosim commit.