Fix issue with extra space that breaks wildcard queries.
This fixes issue #371.  Previously, for a query on "m [0-9]" the
script line "query id wildcard  m [0-9]" would be generated, which
returns no results from Simbad.  Now, the line
"query id wildcard m [0-9]" is generated in the case of a wildcard,
which works.  This hack is a quick workaround; the fragile
_args_to_payload() function should be refactored ASAP.