Fix for issue 1327: HEASARC errors too aggressive
In Issue #1327, the HEASARC module was returning an error even though
the query was successful.  This problem occurred because the code is
searching for the word "ERROR" in the returned text, and for some
returns, "ERROR" was included in the text to mean "measurement error",
not "query error".

Keep an eye on this issue in the future, though: there may be other
classes of HEASARC query errors we're not catching now.