core.py: Added unescaping of raw HTML codes to query result parsing
This change uses the Python standard `html` module to unescape any raw HTML
codes that might be returned in the raw query text and convert them to Unicode.
This should prevent such characters expanding from a single Unicode character
to a multi-space HTML code in ASCII and breaking the FixedWidth formatting
used to create an Astropy Table from the query results.