Changed the cache directory for Splatalogue JSON
Updated the `species_lookuptable` function to check for a local cache, and if not, run the scraping routine to grab the Splatalogue species data.

Also included a new kwarg for this function, `recache`, which will force a re-run of the scraping function even if there is a local cache present.