Allow a frame specification in query_region_async
The original `query_region_async` in the `VizierClass` always uses fk5 coordinates. This is OK for circular query, but limits box queries to boxes oriented along the equatorial axes. This change adds a parameter to `query_region_async` to allow the user to specify galactic coordinates: as a consequence, the returned queries will adopt boxes oriented along the galactic axes, as indicated in the ASU specification (and as tested).