Allow non image data to be retrieved
Non image data (e.g. visibilities) will not have a 'cutout_service' in the datalink response but will have an 'async_service'. As this is also present for all image data we should use this unless we are requesting cutouts (which we will add to our astroquery module in the future).

In addition I have fixed a bug where if there were no files that the user had access to it would give an obscure error when they were asked to be staged.