Change recommended config item access in template
Previously the template core module read the values of the configuration
items at class creation and stored them in class attributes. This meant
that changing the configuration items at runtime had no effect.

Now the template implements private properties that return the current
configuration item value, or the class (or instance) attribute value if
the user has changed it. This enables the users to use the full
functionality of the `astropy` configuration system, but users not
familiar with it can still override it entirely through the attributes.