Update configuration item names in template
The template core module allows users to set configuration items in two
ways, either through the `astropy` configuration system or through class
or instance attributes. The attribute names are no longer in capital
letters because the convention is to reserve such names for constants,
and the configuration item names now match the attribute names.