Require wavelength limits in `get_filter_index()`
Previously the default behaviour of `get_filter_index()` from `svo_fps`
was to make the wavelength range as large as possible, but that results
in a timeout. Now the method requires the user to explicitly set the
wavelength limits because it is better to receive an immediate
`TypeError` about missing mandatory arguments than it is to wait a
minute before an obscure timeout occurs.

See also 90677c86b69e7ea33a131d7a7cfc8282cd157f6b