Simplify handling `DummyResponse` in TAP+ tests
`DummyResponse` is used in `esa/jwst` and `gaia` tests for mimicking
server responses. Now the `message` attribute of an instance is
automatically determined by its `status`, whereas previously it had to
be set manually in a separate step. Furthermore, the `status` of a
`DummyResponse` can now be set through `__init__()`.