Allow `gaia` tests to succeed with `astropy` 4.0
One of the effects of
astropy/astropy@e9d13fcf99ba20ea81d033ef4c096010d1f96531 is that
`"_votable_string_dtype"` key together with the corresponding value gets
added to the metadata of columns read from a VOTable. That commit is
included in `astropy` versions 4.1 and later, but is missing from 4.0.
The fixture that generates data for column attribute comparison in
`gaia` tests now takes the `astropy` version into account.