Allow cache tests and remote tests to co-exist
Previously the cache tests permanently replaced
`requests.Session.request`, which caused the following remote tests to
fail. Now the replacement is handled by a test fixture, which ensures it
does not affect any tests that do not want to be affected. Furthermore,
the fixture can automatically change the mocked response after being
queried once, which reduces manual setup needed in the cache tests.