Remove `astroquery.utils.commons.radius_to_unit()`
`radius_to_unit(radius, unit)` was equivalent to
`astropy.coordinates.Angle(radius).to_value(unit)` except that if `unit`
was not specified then it defaulted to degrees. The convenience of not
having to specify degrees was not worth obscuring the simple
`Angle.to_value()` call, so the function has been removed.