Fix `esa.hsa` remote tests
The `esa.hsa` remote tests attempted to compare sorted lists of actual
checksums with the expected sorted lists, but the sorting was done using
`list.sort()` in the `assert`-statement. `list.sort()` sorts the list
in-place and its return value is `None`, so the tests were actually
checking `None == None`. Furthermore, the checksums obtained in the
tests are not reliably reproducible, so now the tests check the endings
of filenames contained in the fetched tar-archives instead.