Skip tests that need `boto3` if it's not installed
Two `mast` remote tests require `boto3`, which is not installed when
installing `astroquery` with the test dependencies. It therefore cannot
be assumed that `boto3` is installed when running the tests. If it isn't
installed then it is better to skip the tests in question than to have
them fail with `ModuleNotFoundError`.